% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get data from URL and use cookie}
\usage{
get_data(url, session_cookie = NULL)
}
\arguments{
\item{url}{character. Either the url or the day, which guesses a plausible
url.}

\item{session_cookie}{character}
}
\value{
character
}
\description{
Get data from URL and use cookie
}
