% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_mac}
\alias{is_mac}
\title{Are you running MacOS?}
\usage{
is_mac()
}
\value{
boolean
}
\description{
\code{is_macos} checks your \code{\link{.Platform}} and \code{\link{Sys.info}} variables to determine whether or not you are running MacOS.
}
