% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_windows}
\alias{is_windows}
\title{Are you running Windows?}
\usage{
is_windows()
}
\value{
boolean
}
\description{
\code{is_windows} checks your \code{\link{.Platform}} variable to determine whether or not you are running Windows.
}
