% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_to_clipboard}
\alias{write_to_clipboard}
\title{Write data to the clipboard.}
\usage{
write_to_clipboard(data)
}
\arguments{
\item{data}{anything}
}
\value{
data (invisibly)
}
\description{
\code{write_to_clipboard} writes the input data to the clipboard.
}
